Add new test CrossFeature2D and it's associated classes.

+++ Matlab/getNoiseEstimationMap.m
@@ -14,7 +14,7 @@ function output = getNoiseEstimationMap (img_path, block_size)
     %III. estimate noise on each blocks
     [img_w img_h] = size(IMG);
     i = 1;
-    output = zeros(img_w/block_size, img_h/block_size);
+    output = zeros(int8(img_w/block_size), int8(img_h/block_size));
     %output = zeros(10, 10);
     fprintf('\noutput size %i %i\n', size(output, 1), size(output, 2));
     m = 1;
